restructured and implemented feedback
